bpfloader: Conditionally load lock contention BPF programs
Update the BPF loader to load and attach the new lock contention
monitoring program, controlled by a feature flag.

This change introduces a BpfFileDesc for bpfLockContention.bpf,
configuring its maps and tracepoint programs for auto-attachment at
boot.

The loading of this BPF program is now conditional on the state of
the load_bpf_lock_contention aconfig flag. This allows the feature
to be enabled or disabled on devices in the field.

Bug: 442017923
Test: boot test
Test: cat /sys/fs/bpf/lock_contention/
         map_bpfLockContention_contention_latency_map
Change-Id: Ifcddda60b159d0febd49485da986087933604674
Signed-off-by: Kalesh Singh <kaleshsingh@google.com>