bpfloader: Add support for auto-attaching programs
Introduce an 'auto_attach' flag to the ProgDesc struct.

When this flag is true, bpfloader attaches the BPF program and pins
the resulting link. This makes the attachment persist across framework
reboots, only being removed when the device itself reboots.

This simplifies client-side logic, as clients no longer need to
manage attachments or handle missed events after a restart. They can
assume the program is always attached and simply read from the maps.

This is particularly useful for collecting statistics that would
otherwise be corrupted by framework restarts and for capturing events
that occur during early boot, before client processes are running.

Bug: 419635765
Bug: 438637519
Test: manual, load and auto attach ag/35064705
Change-Id: Iffcde8972778b9152f4732e65c221eb72523ea34
Signed-off-by: Kalesh Singh <kaleshsingh@google.com>