Introduce download_vmlinux.py script
This script automates the download of vmlinux artifacts from a given
build ID. It determines the kernel version, creates the appropriate
subdirectories for arm64 and x86_64 architectures, and downloads the
respective vmlinux files.

The script simplifies the process of fetching vmlinux artifacts,
reducing manual steps and ensuring a consistent structure.

Bug: 419635765
Bug: 446262729
Bug: 421990777
Test: ./download_vmlinux.py --bid <bid> --bug <bug>
Change-Id: I5ae3b667af93dcfdaa749cc1d3fc1ccc10207a0d
Signed-off-by: Kalesh Singh <kaleshsingh@google.com>